requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
langchain-core==0.3.51
langchain-openai==0.3.12
langgraph==0.3.30
googlenews==1.6.12
openai==1.74.0